bolstered comments. tweaked test naming for consistency. moved IntHolder to TestHelpers
