+++ test/Jcd.Utilities.Test/TestHelpers/IntHolder.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Jcd.Utilities.Test.Validations
+namespace Jcd.Utilities.Test.TestHelpers
 {
    /// <summary>
    /// A helper class that implements IComparable
